pyproject.toml
[project]
name = "rag-research"
version = "0.1.0"
description = "Citation-graph corpus builder for research RAG (SLSN assistant)"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "pypdf>=4.2.0",
  "chromadb>=0.5.0",
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
]

[project.scripts]
corpus-crawl = "corpus_builder.cli:main"
corpus-viz = "corpus_builder.viz:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/corpus_builder"]

[tool.ruff]
line-length = 100
target-version = "py311"
